requirements.txt
pandas
scikit-learn
nltk
networkx
matplotlib
streamlit
anthropic
python-dotenv
